backend/requirements.txt
# Django
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Auth JWT
djangorestframework-simplejwt==5.4.0

# Database
psycopg2-binary==2.9.10

# Cache / Queue
redis==5.2.1
celery==5.4.0
django-redis==5.4.0

# API Docs
drf-spectacular==0.28.0

# Utils
python-decouple==3.8
